// resources/js/types/hr/employee.ts
export interface EmployeeWorkExperienceInfo {
    id: number;
    employee_id: number;

    company_name: string;
    job_title: string;

    start_date: string;
    end_date?: string;
    is_current: boolean;

    responsibilities?: string;
    achievements?: string;
    company_location?: string;

    reference_name?: string;
    reference_contact?: string;

    created_at: string;
    updated_at: string;
}

export interface EmployeeWorkExperienceForm {
    [key: string]: string | boolean | undefined;
    company_name: string;
    job_title: string;
    start_date: string;
    end_date: string;
    is_current: boolean;
    responsibilities: string;
    achievements: string;
    company_location: string;
    reference_name: string;
    reference_contact: string;
}